added comfirmation alert before deleting  a product

+++ src/app/home/home.component.ts
@@ -1,11 +1,4 @@
-import {
-  Component,
-  inject,
-  EventEmitter,
-  Inject,
-  Injectable,
-  Output,
-} from '@angular/core';
+import { Component, inject, EventEmitter, Output } from '@angular/core';
 import { ProductService } from '../product.service';
 import { Products } from '../products';
 
@@ -42,13 +35,16 @@ export class HomeComponent {
   }
 
   async deleteProduct(id: number) {
-    try {
-      await this.productService.deleteProduct(id);
-      console.log('Product deleted successfully');
-      // Additional logic after successfully deleting the product
-    } catch (error) {
-      console.error('Failed to delete product', error);
-      // Additional error handling logic
+    const confirmed = confirm('Are you sure you want to delete this product?');
+    if (confirmed) {
+      try {
+        await this.productService.deleteProduct(id);
+        alert('Product deleted successfully');
+        // Additional logic after successfully deleting the product
+      } catch (error) {
+        alert('Failed to delete product' + ' ' + error);
+        // Additional error handling logic
+      }
     }
   }
 }
